Don't pull Qt into ProgramEnvironment - and thus every single program.

+++ SerialPrograms/Source/NintendoSwitch/Options/TestPathMakerTable.h
@@ -53,8 +53,8 @@ public:
 
     uint16_t wait_ticks = 0;
 
-    int32_t x_axis = 0;
-    int32_t y_axis = 0;
+    uint8_t x_axis = 0;
+    uint8_t y_axis = 0;
 };
 
 class TestPathMakerTableFactory : public EditableTableFactory{
